+++ tests/test_run.py
@@ -1,16 +1,20 @@
 import numpy as np
 import pytest
 
-from stable_baselines3 import A2C, PPO, SAC, TD3, DQN
+from stable_baselines3 import A2C, PPO, SAC, TD3, DQN, DDPG
 from stable_baselines3.common.noise import NormalActionNoise, OrnsteinUhlenbeckActionNoise
 
 normal_action_noise = NormalActionNoise(np.zeros(1), 0.1 * np.ones(1))
 
 
+@pytest.mark.parametrize('model_class', [TD3, DDPG])
 @pytest.mark.parametrize('action_noise', [normal_action_noise, OrnsteinUhlenbeckActionNoise(np.zeros(1), 0.1 * np.ones(1))])
-def test_td3(action_noise):
-    model = TD3('MlpPolicy', 'Pendulum-v0', policy_kwargs=dict(net_arch=[64, 64]),
-                learning_starts=100, verbose=1, create_eval_env=True, action_noise=action_noise)
+def test_deterministic_pg(model_class, action_noise):
+    """
+    Test for DDPG and variants (TD3).
+    """
+    model = model_class('MlpPolicy', 'Pendulum-v0', policy_kwargs=dict(net_arch=[64, 64]),
+                        learning_starts=100, verbose=1, create_eval_env=True, action_noise=action_noise)
     model.learn(total_timesteps=1000, eval_freq=500)
 
 
@@ -40,6 +44,14 @@ def test_sac(ent_coef):
                 learning_starts=100, verbose=1, create_eval_env=True, ent_coef=ent_coef,
                 action_noise=NormalActionNoise(np.zeros(1), np.zeros(1)))
     model.learn(total_timesteps=1000, eval_freq=500)
+
+
+@pytest.mark.parametrize("n_critics", [1, 3])
+def test_n_critics(n_critics):
+    # Test SAC with different number of critics, for TD3, n_critics=1 corresponds to DDPG
+    model = SAC('MlpPolicy', 'Pendulum-v0', policy_kwargs=dict(net_arch=[64, 64], n_critics=n_critics),
+                learning_starts=100, verbose=1)
+    model.learn(total_timesteps=1000)
 
 
 def test_dqn():